requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
pdf2image==1.17.0
openai==1.12.0
pytest==7.4.3
httpx==0.27.0
python-dotenv==1.0.1
pillow==10.2.0
pdf2image